Platform and type changes, path adjustments
Platforms `AnyCPU;x64` added to all projects.
Set `PublishAot` in `CCSPredict.Console.csproj` to `False`.
Added `ProjectReference` to `CCSPredict.ML.csproj`.
Changed file paths in `Program.cs` to absolute paths.
Changed return value of `CalculateDescriptorsAsync` in multiple files from `Dictionary<string, double>` to `Dictionary<string, float>`.
CSV configuration in `CsvCcsDataProvider.cs` adjusted.
New properties `MZ` and `Adduct` added to `MoleculeWithCcs`.
Updated project definitions and configurations in `CCSPredict.sln`.

+++ CCSPredict/CCSPredict.Core/Interfaces/IMoleculeDescriptorCalculator.cs
@@ -4,6 +4,6 @@ namespace CCSPredict.Core.Interfaces;
 
 public interface IMoleculeDescriptorCalculator
 {
-    Task<Dictionary<string, double>> CalculateDescriptorsAsync(Molecule molecule);
+    Task<Dictionary<string, float>> CalculateDescriptorsAsync(Molecule molecule);
     IEnumerable<string> SupportedDescriptors { get; }
 }
